+++ test/highlight_test.rb
@@ -32,7 +32,7 @@ class HighlightTest < Minitest::Test
 
   def test_field_options
     store_names ["Two Door Cinema Club are a Northern Irish indie rock band"]
-    fragment_size = ENV["MATCH"] == "word_start" ? 26 : 20
+    fragment_size = ENV["MATCH"] == "word_start" ? 26 : 21
     assert_equal "Two Door <em>Cinema</em> Club are", Product.search("cinema", fields: [:name], highlight: {fields: {name: {fragment_size: fragment_size}}}).first.search_highlights[:name]
   end